move db to separate namespace
